Updated configure.in for proper debug and optimization handling.
The new configure script now defaults to no debugging; if you use the
--enable-debug option to configure, it uses "-g" for debug.

The optimizer section applies the compiler warning options and optimizer
options separately.  If debug is requested, no -O options are added.

For IRIX 6.2 and higher, and for GCC/EGCS, -O3 is used for optimization.
For IRIX 5.x -O2 is used.  For HP-UX +O2 is used.


git-svn-id: file:///fltk/svn/trunk@261 ea41ed52-d2ee-0310-a9c1-e6b18d33e121