Added
   if (parent_) ((Fl_Group)parent_)remove(this);

to Fl_Widget's destructor which helps keeps a FLTK app from crashing if
a widget is destroyed before being properly removed from its parent.
This should not add much overhead because normally the widget should
already be remove()ed frum the parent by the time the destructor is called
and parent_ will therefore be 0.  However, in some useful circumtances it
is easier for the widget to be destroyed before being removed:

   class Combo : public Fl_Group {
   public:
     Fl_Button bt;

     Combo(int x, int y, int w, int h, const char* l) :
       Fl_Group(x, y, w, h, l), bt(x, y, w, h, l) {}
   };

the Fl_Button (bt) will always be destroyed before the Fl_Group.

In fact, this is the way Fl_Color_Chooser works.  The only reason it doesn't
crash when it is used is because usually nothing needs to access the group
between the time the widget is destroyed and the time the parent group is
destroyed.  However, it is not a good idea to assume that this will always
be true because the widget's destructor may need to do something that needs
to access the group (like throw the focus).

-Carl


git-svn-id: file:///fltk/svn/trunk@607 ea41ed52-d2ee-0310-a9c1-e6b18d33e121