Cleaned up and sped up the Fl_Group code (suggestions from Curtis)
Removed the hack that saved a malloc for 1 child of a group.  This
allows Fl_Group::array() to be an inline method, which probably saves
more space than the hack did.

Simplified Fl_Valuator by making step_ a single value.

Cleaned up the code for Fl_Slider a lot.

Fl_Slider::slider_size() is now an integer number of pixels.  This may
be incompatable but makes it possible to draw fixed-sized glyphs as
the slider.  I added a method slider_position(W,S) so that a subclass
can find out the pixel position of the slider in order to draw it.

Fl_Scrollbar now has a minimum slider_size of a square, and clicking
outside the slider causes page up/down.


git-svn-id: file:///fltk/svn/trunk@757 ea41ed52-d2ee-0310-a9c1-e6b18d33e121