Rewrote the glpuzzle demo program to not use glut and put this in the test directory (the original still remains in the glut directory)
Fl_Style uses the existence of the revert function to determine if
they are local or not, got rid of the "dynamic" member.

Put the inactive boxtypes code back, but this time I fixed the code so
it produces the same colors (rounded to the nearest entry in the
colormap) as Carl's code does.  The output looks identical to me.
Also restored the box edge patterns to the less-motif and more rounded
style I was using (not sure if this was intended by Carl or a
mistake).


git-svn-id: file:///fltk/svn/trunk@850 ea41ed52-d2ee-0310-a9c1-e6b18d33e121