Fixed the fonts demo to show characters with the high bit set (including the "hole" that MicroSoft filled in on their fonts)
I want the default appearance to be *identical* on all platforms.  If
this means changing it to be more windoze-like for people to accept
the toolkit as-is, then I'll do that.  Current differences from win98
appearance are:

1. The black triangles in menus and scrollbar buttons are larger.

2. Vertical spacing between menu items is much smaller.

3. Highlighting of menubar is disabled, to be consistent with the fact
   that highlight is disabled everywhere else.

4. Check boxes and radio buttons in menus look just like check and
   radio buttons on the main windows.

5. The shading of button borders were selected to look more rounded and
   friendly, and they do match *some* Win32 applications.

Changes required:

   Made FL_NORMAL_SIZE be 12 (it is recommended you shrink all your
   widgets to match, for instance menubars should be 25 pixels tall).

   Made FL_HIGHLIGHT_UP/DOWN_BOX draw a THIN box.  We do not seem to be
   using the FL_FLAT_UP/DOWN_BOX and maybe these should be deleted.

Carl: if the user changes the colors on windoze it will wipe out any
loaded theme.  I recommend you split the color-setting portion from
the rest of fl_windows() and put it as a static function into
Fl_win32.  Call it when the first window is shown and in response to
the event.  Don't call fl_windows() anywhere by default.  Also remove
the duplicate code from Fl::arg().


git-svn-id: file:///fltk/svn/trunk@876 ea41ed52-d2ee-0310-a9c1-e6b18d33e121