Made FL_UP_BOX and FL_NORMAL_BOX be the same again.  The change broke several old programs that hard-coded FL_UP_BOX in and assummed it went up/down.
The calls to get the values for FL_VIEWCHANGE events are changed to
simple inline functions called Fl::event_dy(), etc.  Removed the "mode".

Removed the configuration for the wheel.  My main concern is to keep all
calls to "getconf" in a single source file so this can be rewritten to a
straight parser.  Also I don't think these are needed.

The wheel now reports dy in "pixels", one click is 12 pixels (this appears
to be what Microsoft intends, as many other parts of the WIN32 gui measure
things in 10ths of a pixel).

Browser changed to have the wheel move it by "pixels" and to do something
rudimentarily intelligent if they move it while holding a button down.

Rearranged and simplified the Fl.cxx and Fl_Group.cxx event handling
so the new FL_MOUSEWHEEL events are handled by the same case as
several other events, and hopefully new events will also go there.


git-svn-id: file:///fltk/svn/trunk@926 ea41ed52-d2ee-0310-a9c1-e6b18d33e121