- Updated VC project partly (core and images only). Defined an FL_IMAGES_API symbol for the images DLL. Fluid works using the DLL library (it uses two DLL, fltkdll and fltkdll_images). Fluid plugins work with VC.
- I've done the following for the TrackMouseEvent problem :
I'm using LoadLibrary and GetProcAddress to detect if the TrackMouseEvent function is
present in User32.dll.
I'm using this pointer to make the call. Additionally, I test for _TrackMouseEvent in
COMCTL32.dll (under WIN95, TrackMouseEvent will never be defined, only _TrackMouseEvent)
If all this fail, the call is not performed, but FLTK should still work (I'm not sure, it
seems I have _TrackMouseEvent on my system, so I could not test without it)


git-svn-id: file:///fltk/svn/trunk@937 ea41ed52-d2ee-0310-a9c1-e6b18d33e121