Added Fl.H required by an inline function if Fl_Repeat_Button.H
Fl_add_idle adds new functions to the end of the queue ring, rather than the
start, so they are executed in the order added, and a callback that adds
itself does not prevent others from being called.

Fluid lets you type in code that starts with '#' for cpp directives.


git-svn-id: file:///fltk/svn/branches/branch-1.0@969 ea41ed52-d2ee-0310-a9c1-e6b18d33e121