add_idle adds the new callback to the end of the idle list, rather than the start, as suggested by Cliff Miller.  Allows a callback to remove/add itself.
Fixed the redraw of the huge scrolling menu (it was not testing for
the FL_EXPOSE_DAMAGE bit)

Fixed the bad redraw of the checkmark in the menu of the menubar demo
(caused by it using the wrong color for a highlighted menu title if
the highlight color was zero)

Removed the deactivate() method from Fl_Repeat_Button, this inline
function required Fl.H to be included.  Not clear what it was doing
or if it is used, you can get the same effect by calling the normal
deactivate().


git-svn-id: file:///fltk/svn/trunk@970 ea41ed52-d2ee-0310-a9c1-e6b18d33e121