Renamed fl_set_fonts.cxx to fl_list_fonts.cxx (since that is the name of the function it defines).  Added a TOTALLY UNTESTED fl_list_fonts() to the win32 version.
Added fl_font() and fl_size() that return the most recent values
passed to fl_font(font,size).  These can be used to save/restore the
font or to check which one is in use in labeltype functions, and to
avoid having to track it yourself in complex drawing operations.

Patches from Alexander Mai for OS2, mostly to the file chooser
drive-listing code.

The glyph() function is now responsible for changing the color of the
label in response to FL_INACTIVE, the color passed to it is normal.
This matches how label and box drawing functions work, allows a lot
more freedom to design the glyph functions, eliminated some code, and
I found lots of code assummed this anyway...


git-svn-id: file:///fltk/svn/trunk@976 ea41ed52-d2ee-0310-a9c1-e6b18d33e121