Fixed the arrow and page up/down direction on vertical scrollbars.
Lots of fixes to clean up Fl_Scroll.  The scrollbars are no longer
visible children, so calling clear() and children() and other things
on an Fl_Scroll do what is expected.  The redrawing is fixed to use
the same type of clip-out as the group for nice non-blinking resizing.
They scroll automatically to show the widget with the focus.  This
required much rewriting and cleanup of Fl_Group code and event
handling.

Changed how scrollbars draw so they don't blink either.  Requires a
kludge in motif-style themes to draw the background behind arrow
triangles.

Renamed Fl_Widget::clear_damage(n) to set_damage(n) for obvious
clarity reasons.  clear_damage() with no argument remains (it does
set_damage(0))

New call Fl_Widget::relayout() that sets the layout flag.

FL_FOCUS events are sent to all parent widgets whenever a child gets
the focus, even if it is from another child.  This allows the parents
to track which child has focus (and the Fl_Scroll to scroll to show
it).

Changed return value of fl_clip_box to be the same as fl_not_clipped,
and implemented partial-clip detection on Win32.  This was needed to
get scroll to draw correctly when both scrolling and a child being
damaged happened at the same time.

FL_MOVE events are sent to all widgets even while the mouse is pushed,
as CET requested.  The highlight now follows the events, and it can
allow drag&drop interfaces.  Disabled while menus are up.


git-svn-id: file:///fltk/svn/trunk@1060 ea41ed52-d2ee-0310-a9c1-e6b18d33e121