Made colors for glyphs even better.
glyph background is selection_color when FL_VALUE is true - sometimes you
want selected glyphs to be a different color!

glyph foreground is text_color - less likely to cause collisions than
label_color.  Selected version is selected_text_color

AFAIK, doing this avoids color collisions with all of the current widgets.
The one special case are the "nice" sliders (of course), where selection_color
is used for the color of the color mark.  This special case is desirable
because "nice" sliders can be used in value outputs, which already use the
text colors.

I played around with this for a while, and this setup seems to work "better"
for all cases.

Carl


git-svn-id: file:///fltk/svn/trunk@1067 ea41ed52-d2ee-0310-a9c1-e6b18d33e121