Moved Fl_Grab stuff to the main Fl.cxx to get rid of public static variables.  Moved Fl_Window:: methods to Fl_Window.cxx to make them easier to find.
Fl_Output takes the focus when you click on it (so there is only one
selection visible).

Fixed popup menu titles to use an FL_UP_BOX if the window_box of the
button is FL_NO_BOX.

Since selection_color() and selection_text_color() can no longer be
zero, I removed the tests for that.

Removed the now unused FL_CLOSE event (the code directly calls the
window's callback rather than use this, and I thought it best to
delete the symbol so noone thinks it does something).  I moved
FL_KEYUP to fill in the enumeration, because it puts it next to
FL_KEY, so be sure to recompile!

Fixed the scrollbar in the valuators demo (it was never initialized so
the minimum (bottom) was zero, and maximum (top) was 1).  In normal
use scrollbars are always initialized.

Reversed the vertical roller just like the slider, so that the minimum
is at the bottom.  Also fixed the shortcut keys so they are just like
sliders.

I patched the crash in doublebuffer, but this is still wrong and I
need to fix it.


git-svn-id: file:///fltk/svn/trunk@1073 ea41ed52-d2ee-0310-a9c1-e6b18d33e121