Added Fl_Widget::width() and Fl_Widget::height(), these are the same as w() and h() but call layout() if necessary.
Added an X version of fl_get_system_colors that matches the fltk 1.0
version.  It reads the X resource database and understands attributes
set by the krdb program in the KDE destkop.  You can also give
app-specific colors by using foo.background instead of *background in
the database.

Renamed window_color to text_background and window_box to text_box.
These names seem to make more sense and match what X/kde used.

Fl::grab(window*) does not call show() on the window.  Changed it to
accept a widget instead of a window.

To fix the highlight blinking I left the group as it was and instead
changed all the widgets that accept FL_ENTER to return 1 for FL_MOVE
as well (most already did this).  This will allow a widget to be
non-rectangular by returning 0 if the cursor moves out of it, I
thought this may be a useful addition, even though it was discovered
by accident.

Some fixes to the fonts and colbrowser demo.  After much debating I
decided to leave the value() of browser starting at zero and require
the application programs to be fixed.


git-svn-id: file:///fltk/svn/trunk@1145 ea41ed52-d2ee-0310-a9c1-e6b18d33e121