Various bugfixes.
KDE theme now handles both KDE1 and KDE2.  I tested it with the KDE2 beta,
can someone who has KDE1 try it to make sure it still works?  The algorithm
to decide whether to do KDE1 or KDE2 is pretty lame.  If the file
$HOME/.kde/share/config/kdeglobals exists, it assumes you are using KDE2.
Can anyone think of a better way to know which version of KDE is being run?

Carl


git-svn-id: file:///fltk/svn/trunk@1146 ea41ed52-d2ee-0310-a9c1-e6b18d33e121