Same changes as to 1.0 for simplifing the timeout/wait code.
Fl::wait(time) returns the result of calling select(): negative for an
error, 0 if nothing happens, and >0 if anything does (on Win32 0 is
returned for at timeout of 0.0 and 1 for anything else).

Fl::wait(void) returns the same value (and in fact just calls
Fl::wait(time) with an "infinite" time).  This return value is
incompatable with 1.0.

Fl::check(void) returns the same value (and in fact just calls
wait(0.0)).  This is also incompatable with 1.0.

Replaced Carl's calls to atol() with strtol or strtoul.  These
functions are used elsewhere in fltk and appear to be portable.
Deleted the replacement atol code.

Added Fl::has_timeout and Fl::has_idle.

Fixed bug with multiple child windows (like the cube demo).
FL_SHOW/HIDE was only sent to one of the children.


git-svn-id: file:///fltk/svn/trunk@1216 ea41ed52-d2ee-0310-a9c1-e6b18d33e121