Slight cleanup of the Win32 drawing code, I deleted the fl_current_xmap structure and put the rgb, pen, and brush in single variables.  This matches how most of the rest of fltk's drawing state is stored.
Made fl_draw_image work on 8-bit StaticColor visuals.  It is not ideal
but these are rarely (if ever) used.

Change the private style_win Carl added to a public called
fl_message_window, and made the cut&paste, and fl_kde() use this same
window.  Probably there are other places that can use it as well.  I
wasted some time trying to get grab() to use it, apparently X requires
the grab window to be mapped.  Carl: We may want the atom to be
FLTK_MESSAGE_WINDOW, and change the message to be FLTKChangeStyle so
that the names match kde more closely.

Fixed the Alt shortcut for the menubar.

Replaced the call to getconf_sections with interation through all the
named styles.

Added the GL_SWAP_TYPE stuff.


git-svn-id: file:///fltk/svn/trunk@1249 ea41ed52-d2ee-0310-a9c1-e6b18d33e121