Xdnd works in both directions now! It can only do text, and the only widget that uses it is Fl_Input. Data types and actions are currenlty not supported, and I don't think they can be supported without system-specific code. I hope to make the necessary variables public in the Fl_x header files with enough hooks that it is possible to make an X file manager, or to drag or accept non-text, with system-specific code.
Many fixes for SelectionNotify copied from Jean-Marc's dnd code.

Deleted Fl_Input_ and merged everything into Fl_Input. The base class
was pretty much useless.

Instead I added new virtual method replace(...) to Fl_Input. This is
called to make any changes. Subclasses can override this function to
limit the editing that can be done. This made implementation of
Fl_Output and Fl_Float_Input much simpler!

Fixes made in 1.0 for wordwrap have been merged in, and wordwrap is
available by using the new Fl_Wordwrap_Input (or _Output)
widgets. This does not replace the need for a fancy text editor, as it
only works well for small portions of text and there are no
scrollbars. If nedit works well enough it may be preferrable to delete
the multiline/wordwrap handling to make Fl_Input much simpler.

Deleted selection_owner, which was not being used and confused the dnd
stuff.

I tried to get Fl_Text_Editor to do fltk's compose sequences. Seems to
work, but this bypasses some of the indirection tables so it may break
the ability to reprogram the printing characters.

PS: nedit seems to work really well!


git-svn-id: file:///fltk/svn/trunk@1272 ea41ed52-d2ee-0310-a9c1-e6b18d33e121