fl_modal_for had problems:   1. Creating a non modal window then deleting it then creating another      window crashed the application.   2. Creating a non modal window made all subsequently created windows      non modal. Added modal_for_ pointer to Fl_Window so each window keeps track of its modal_for_ pointer itself.
Fixed my last fix which allowed windows to actually know where they are
on the screen.  The proper time to do this is when the window is mapped,
not when it is created.  This allows windows to be placed by the window
manager when no X or Y coordinates are specified which was broken by my
last "fix."

Added public Fl_Text_Buffer::call_modify_callbacks(void) which will call
the modify callbacks associated with a buffer with nothing changed.

Carl


git-svn-id: file:///fltk/svn/trunk@1274 ea41ed52-d2ee-0310-a9c1-e6b18d33e121