More fun with modal_for!  Now should never, ever cause crashes(?) and as an added bonus actually should always come up with the correct window to be modal for.
This will (and should) crash, however:
  Fl_Window* w = new Fl_Window(100, 100);
  Fl_Window* m = new Fl_Window(100, 100);
  m->set_non_modal();
  m->modal_for(w); // force window m to always be subordinate to window w
  delete w; // uh-oh
  m->show(); // will crash here

The moral of the story is that if you want to force a window to be
subordinate to another window using the Fl_Window::modal_for(Fl_Window*)
then you need to make sure that the Fl_Window it is subordinate to still
exists when the subordinate window is shown.

If you don't force a window with Fl_Window::modal_for(Fl_Window*) (nothing
does right now except the editor demo), then it will use Fl::first_window()
as it always has in the past (but will never crash and will always use the
right window unlike in the past).

Carl


git-svn-id: file:///fltk/svn/trunk@1280 ea41ed52-d2ee-0310-a9c1-e6b18d33e121