Minor bugfixes to get themes to compile again (though they don't seem to be working properly with Bill's latest styles change).  The forms library is still broken.
Combined handling of X ConfigureNotify and MapNotify events which were
handled the same way.

Tooltips now calculate geometry in layout() as they should.

Hiding an Fl_Window destroys the window instead of unmapping it because
at least on my X system just unmapping it causes undesired events to be
generated.  For example, moving the cursor over the area of an unmapped
window causes the window behind it to get a LeaveNotify event!  Not sure
whether this is a bug in XFree86 4 or not...  This was the main problem
with tooltips.

Code clarity changes to modal_for stuff.  Functionally the same.

Carl


git-svn-id: file:///fltk/svn/trunk@1282 ea41ed52-d2ee-0310-a9c1-e6b18d33e121