Removed the public Fl_Input/Output::default_style, which were not being used (where they were used it had no effect as the widget already had that style).
Renamed fl_shortcut_label to Fl::key_name. Renamed the source file
fl_shortcut to Fl_key_name to match.

Documentation fixes for my changes to the methods used to draw things.

Got rid of a lot of warnings when it is compiled with -Wwrite-strings.

Got the fractals demo to work again.

Removed the Emacs control characters from Fl_Input and Fl_Text_Input,
and put a fallback translator to arrow keys in Fl.cxx. They will thus
work as long as nobody redefines the control characters on a
menu. Added the symbol Fl_Clear for ^K, this should do a clear-to-eol.


git-svn-id: file:///fltk/svn/trunk@1291 ea41ed52-d2ee-0310-a9c1-e6b18d33e121