Changes to the fl_vertex drawing code to be more PostScript like. It now has only a begin(), and fl_end_line(), fl_end_polygon(), etc decide what to do with the path, and the path is saved. This allows you to fill and outlike a polygon with only one description of the path. fl_gap() now works for fl_end_line() as well.
Makefile changes: "make" just makes the static version.  However
"make install" and "make shared" will build the shared library.
I also made the top level makefile exit if there is an error running
make in any of the subdirectories. These help a lot for building test
versions of fltk.

I tried making Fl::run() run forever and having the window callback
do exit(0) when the last window closes. This should make no difference
for any program that keeps one window open and seems to be easier to
explain.

Some cleanup of the back-compatability portion of Fl_Window::modal_for().
The code is now all in Fl.cxx rather than the machine-specific part.
(warning: this required an untested change to the Fl_win32 code)

Fl_Window::hide() does not destroy the window, just unmaps
it. Destruction of the window, changing the modal_for(), or calling
destroy() actually gets rid of the window. I am prepared to put this
back, as there are warnings it causes problems.

Lots of fixes to the documentation.


git-svn-id: file:///fltk/svn/trunk@1307 ea41ed52-d2ee-0310-a9c1-e6b18d33e121