The projects couldn't build under Win32 in Visual Studio 6.  Mostly I had to change preprocessor definitions (FL_API, FL_DLL, FL_SHARED, etc.) to get it to compile.  Also had to change what files are in the dll and library projects, since they were referencing stuff that no longer exists.
This *does* build from a clean start on Win32 under Visual Studio 6 now.  It
does *not* work correctly.  Virtually all of the demo programs have problems,
the most obvious of which is that button labels aren't drawn correctly.  (The
labels are entirely invisible, so I'm not sure without digging a little deeper
just where they're going...)

For the moment, everything is still in a single library in the VS6 project --
the fltk_images, fltk_gl, and so on are not yet broken out into their own
projects.

I also have not yet double-checked that this builds without error under GCC,
but will do so later today or tomorrow.  In the meantime, just slap me if I
broke anything.


git-svn-id: file:///fltk/svn/trunk@1372 ea41ed52-d2ee-0310-a9c1-e6b18d33e121