Fixed text output on Win32.  Wasn't considering positions correctly (needed to add fl_x_ and fl_y_ to offset).
Fixed subwindow positioning in *nix (similar problem, wasn't using offset from
parent container -- was most obvious in "tile" and "subwindow" demos).

Partial fix to subwindows in Win32.  (Wasn't considering border size correctly
when the window was embedded in another widget.  I just set the border offsets
to all zeros if parent() is non-NULL.)

Problem still exists with subwindows in Win32.  Try the "tile" demo for the
most dramatic example of the problem.  (The problem is that every time you
cause the window position to be recalculated, it adds the offset from
the parent, cumulatively to the previous offsets.  So your subwindow rapidly
walks off the screen.  I tried a few different ways to stop it, but I seemed to
wind up with one that didn't move at all, or would move/size correctly but not
line up in the tile the way it was supposed to.  This is only a problem on
Win32 -- the *nix build works fine.)


git-svn-id: file:///fltk/svn/trunk@1373 ea41ed52-d2ee-0310-a9c1-e6b18d33e121