Added a workaround for a problem with tooltips in Win32.
The problem is that after a tooltip has been displayed the first time, any
subsequent tooltips are shown active, causing the active window to lose
focus.  This despite the fact that the Win32 ShowWindow() *is* being called
with the SQ_SHOWNOACTIVATE flag set.

The workaround I put in is that on Win32, when the tooltip exits, it deletes
the tooltip window, causing a new tooltip window to be used each time, rather
than reusing the same one.  Obviously this isn't the most desirable way to
handle this, but it does seem to work.  (Note that this only effects Win32
builds.)


git-svn-id: file:///fltk/svn/trunk@1385 ea41ed52-d2ee-0310-a9c1-e6b18d33e121