Added multiple column capability to Fl_Browser.  This was present in FLTK 1.x, and is in the documentation for 2.x, but just plain wasn't there.  Because of the fact that Fl_Browser is now just a glorifed Fl_Menu_, the functionality for multiple columns was actually put into Fl_Menu_.  The column_char() functions from Fl_Browser were moved to Fl_Menu_ instead.   The column_widths() functions were added -- note that they just copy the pointer, *not* the array of widths. Functions of the same names were also added to Fl_Item, since it has to know about them to do the drawing.  (Again, only pointers are copied).
I have not yet changed the documentation -- figured I'd wait and see if this
was going to cause any trouble first.  I ran through all the demo programs that
use menus and didn't see anything amiss.  I also tried a multi-column browser
here (didn't add to CVS, just tested locally) and it looks good too.

If this was a silly way to do it, or if there's a better way, please enlighten
me.  But we do need this functionality here...


git-svn-id: file:///fltk/svn/trunk@1419 ea41ed52-d2ee-0310-a9c1-e6b18d33e121