Minor correction in the file chooser, and a minor change.
Correction was that if you're in a large directory listing and scrolled down in
the list, then move to a different directory, the list of files doesn't
reposition itself.  So if the new directory only has a few files, the list may
appear to be empty, but it isn't, it's just scrolled completely out of view.
Corrected.

Change was to add an alternate interface to the fl_file_chooser function call
so that you can specify an additional argument, a boolean indicating whether
the call is for a save dialog (true) or a load dialog (false).  Only has any
meaning under Win32 at the moment.  I put this in as an overloaded function
instead of just using a default argument on the existing interface, so that
it's still binary compatible and you don't have to recompile everything.


git-svn-id: file:///fltk/svn/trunk@1422 ea41ed52-d2ee-0310-a9c1-e6b18d33e121