Corrected the problem with the default font coming up bold on some Windows systems.
I'm not sure whose the problem is, but what happens is that when
enumerating the fonts, Windows reports some of the fonts only with their
bold name.  For example, on my test system, I have the default font set to
"Tahoma".  There is a "Tahoma" font and a "Tahoma Bold" font on my system.
When Windows enumerates the fonts, it only reports "Tahoma Bold" and not
"Tahoma".  (There probably is some reason it does this, but I don't know
offhand what it might be.)  So when FLTK generates the fonts, the only
versions of "Tahoma" all come out bold (since they're based on the bold
version of the font).

So all I did was have FLTK watch when enumerating the fonts, and if it gets a
font which ends in an attibute (ie "Tahoma Bold", "Arial Italic" or the like),
FLTK will strip the attribute off the name before creating the font.  On all
the fonts on my system, this works just fine.  (Note that this change was
made only in the Windows code.  Other platforms are unchanged, but I've never
seen the problem anywhere but Windows.)


git-svn-id: file:///fltk/svn/trunk@1511 ea41ed52-d2ee-0310-a9c1-e6b18d33e121