Keyboard focus box of Fl_Light_Button draws black rather than the light color.
To make the focus go to Fl_Input widgets when the window is first
shown, they now return 2 for handle(FL_FOCUS) and groups will search
for these and put the focus on them (if it fails it puts the focus on
the *last* widget, which matches how MSoft dialog boxes work). To make
this work fully all subclasses of Fl_Group must return 2 for FL_FOCUS
if thier children do, this is only done for Fl_Group and
Fl_Tabs. However I may change how this is done.

Does not call draw() or layout() on zero-sized windows. Destroys any
region accumulated for these or invisible windows.

Tried to redesign the web pages somewhat. Feel free to criticize, or
try your own design.

Changed shortcuts in fluid to use Ctrl instead of Alt.


git-svn-id: file:///fltk/svn/trunk@1586 ea41ed52-d2ee-0310-a9c1-e6b18d33e121