Fl::belowmouse() was resetting the tooltip, causing the flashing in the FLUID window and havoc for the KDE 2.2 window manager...
Changed the default tooltip position to be +10, +10 from the
mouse pointer, and the default tooltip background color to a
pale yellow to match other toolkit defaults.

Update Fl_Widget::flags_ to be an integer, to support more flag
bits.

Update Fl_Window to support override() and set_override() methods.

The configure script did not set the default library path to
/usr/lib32 under IRIX 6.2 and higher when prefix is /usr.


git-svn-id: file:///fltk/svn/branches/branch-1.1@1637 ea41ed52-d2ee-0310-a9c1-e6b18d33e121