Xlib font selection fixed up for changing encodings. Before if an encoding was not found for a font it got marked so that no encoding search would be done, even if you switched to an encoding that exists.
If you set the value() of a browser that has not been shown yet or has
had it's contents changed, it will center the item. It used to scroll
to incorrect positions.

"make depend" makes dependencies for the .shared.o files as well.

fl_encoding(x) does not change current font, instead it changes the
next fl_font(x) call.

Made Fl_FontSize object private to the various fl_font
implementations. This should make it easier to support printing models
where there is no need to cache every font+size we ever found.

On X, if a child window completely filled the parent it would prevent
expose events from being sent to the parent, defeating the parts of
fltk that rely on this for synchronization with the window
manager. One known effect is that if you call resize() on the window
it would appear to be iconized() and refuse to draw() or layout() from
then on.


git-svn-id: file:///fltk/svn/trunk@1685 ea41ed52-d2ee-0310-a9c1-e6b18d33e121