X sent fake keyup events when you hold down a repeating key, made fltk ignore these.
Typing a key while a menu was up that did not match a menu item shortcut
crashed.

Fl_Input and Fl_Text_Editor now do a bunch of Emacs Alt+letter
combinations. These (and Emacs ctrl+letter keys) are only done if the
program does not use them for a shortcut such as a menu item. The
implementation of this cleaned up and removed from Fl.cxx.


git-svn-id: file:///fltk/svn/trunk@1946 ea41ed52-d2ee-0310-a9c1-e6b18d33e121