fl_background() will produce usable gray shades when the color is white or black (makes the high-contrast Windows styles work).
Lots of fixes to the WIN32 drawing code to get rid of blinking and
make the windows update when the application is not active.

Greg's rotation slider added to the symbols demo.

Added Fl_Multi_Image and multiimage demo program.

Valuator reported step() as Infinity, which messed up fluid's saved
files. Added ability to set line & page size in fluid.


git-svn-id: file:///fltk/svn/trunk@2198 ea41ed52-d2ee-0310-a9c1-e6b18d33e121