Fixes to the colors to make it easier to change themes:
Buttons use the buttonbox() and buttoncolor() to draw unless the box()
or color() are set directly on them. This gets rid of most of the need
for subclasses to have non-empty default styles, so themes can set
almost everything they want by modifying Widget::default_style.

Items in a browser or menus will use the textcolor() of the parent
browser/menu to draw unless labelcolor() is set directly on them. This
is compatable with 1.1, and allows items to appear differently in
menus than in browsers.

Tooltips use textcolor rather than labelcolor for the text. This seems
to match what KDE does.

The KDE and Win32 theme readers fixed to take advantage of above changes.

Buttons that have NO_BOX and label()==0 are drawn specially on the
assumption that the image() is a special animating one that shows the
current state.

Added color argument to Image::draw(). This is only used by bitmaps to
pick the color to draw, but I suppose other image types might use
it. This also makes the method the same as the method for labels and
symbols.


git-svn-id: file:///fltk/svn/trunk@2938 ea41ed52-d2ee-0310-a9c1-e6b18d33e121