If X_HAVE_UTF8_STRING is not defined, then the X implementation lacks Xutf8LookupString() and is probably not XFree86. XmbLookupString(3X11) is used in its place and returns a locale-dependent multibyte-encoded buffer which is trancoded to UTF-8 via iconv(3). If iconv_open() does not find a conversion description for multibyte to UTF-8, the non-XInput method is used.
Additionally, if the configure script cannot find Xutf8LookupString() in libX11, it attempts to find iconv and defines the symbol HAVE_ICONV_H on success. If iconv is provided by libiconv (rather than libc), '-liconv' is added to the linker LIBS.

This seriously uglifies the KeyPress handler code, so a more elegant solution for X lacking UTF-8 facilities is sought. The 'utf' test does seem to work, but I am unsure how to really exercise this stuff. Note that this fix could be applied to the 1.2.x branch per STR #438.


git-svn-id: file:///fltk/svn/trunk@3468 ea41ed52-d2ee-0310-a9c1-e6b18d33e121