More patches from Yasuhiro Matsumoto:
0. Code to draw X fonts by selecting between two fonts for each letter. I
   put the code in but commented it out because it did not work on my
   Irix machine and it makes all text do an X call for each letter.

1. added the code for dynamic loading IMM32.dll for IME.
   (some environment don't have imm32.dll)

2. bug for IME position
   (fix duplicate transform in Fl_win32.cxx)

3. set spot only if focused().

4. unset spot if widget leave focus.
   (if widget is not Fl_Text_Display or Fl_Input, stop input method)

5. enable XIM on DBCS(double byte character set) (I think this is a
   change made to Windows version)

Also he fixed the old color chooser popup to stay in place!

I also rearranged xutf8.cxx some so that if iconv does not work it
runs the code written for !iconv. However there is no real sign that
this works, it looks like if keysyms are to be used, the input method
has to be turned off, because it just eats them and we never see
them. Another possibility is to use XwcLookupString and just assumme
it is returning ISO-10646-1.


git-svn-id: file:///fltk/svn/trunk@3476 ea41ed52-d2ee-0310-a9c1-e6b18d33e121