Quartz port: - removed all #warning statemnets and explained caveats with regular   // warning lines - made GL text rendering work again (may use wrong fonts since agl doesn't   really support ATS) - removed horizontal jittering in Fl_Input_ caused by fractional text   positions. Unfortunatly, this does not work for Fl_Text_Display, which   not only jitters, but also became horendously slow due to the complex   font rendering in Quartz)
I will fix some of the speed issues in fl_measure this afternoon, and
also improve on the 'innards' for rgb data rendering. After that, I will
have to leave the Quartz port alone for a while due to lack of time.
I suggest we make another rc in the next days.


git-svn-id: file:///fltk/svn/branches/branch-1.1@3812 ea41ed52-d2ee-0310-a9c1-e6b18d33e121