Support for boxes that change their insets depending on style and flags, used this to make buttons that move when you push them.
NumericInput messed up on the first up/down arrow or mouse wheel.

NumericInput only responds to wheel if it has focus.

Wheel events sent to focus if belowmouse does not want them.

Subwindows are drawn like regular widgets, in the middle of updating
their parent. This reduced the complexity of the OS/X version by removing
the need to clip to them. New GSave object to allow you to draw another
surface and then return to the current one.


git-svn-id: file:///fltk/svn/trunk@4245 ea41ed52-d2ee-0310-a9c1-e6b18d33e121