STR #831: When moving an OpenGL window outside the screen and then back again lets fltk get stuck.
The problem seems to be an idefinite repeat of the WM_PAINT message. The driver sends an internale WM_PAINT (one that does not contain an update region). FLTK then creates a union with the still existin 'i->region' of the OpenGL window, causing yet another WM_PAINT message.

The fix is to not run through the WM_PAINT handler at all if there is no update region (as suggested by Microsoft). This change is definetly debatable, since it does modify some internal FLTK behavior. I will commit it anyways for now, but we should pay very close attention to redraw problems by those who use the svn releases.

git-svn-id: file:///fltk/svn/branches/branch-1.1@4456 ea41ed52-d2ee-0310-a9c1-e6b18d33e121