Fixed STR #309 The important radio item unworking (uncoded) feature has been coded.
Furthermore, I added genericity to the radio+toggle visual types
and unified the button enum, item enum,  and code to the Widget level.
Then I removed the redondancy found in Button and Item (same enums)

So now not only items and buttons can use setonly() and gliph features 
but potentially any widget in a group having a type to RADIO or TOGGLE.

This should open the door to other uses of 'reset all but one' 
characteristics to other widget than item and buttons.

I really LOVE fltk2 ;-)



git-svn-id: file:///fltk/svn/trunk@4931 ea41ed52-d2ee-0310-a9c1-e6b18d33e121