Made most of the .fl files used by fluid work again to produce the code of fluid. Mostly I broke them by taking code that was fluid-only and removing the fltk namespace from it.
Two problems:

First the about panel cannot be compiled as the source image is missing.
And fluid insists on inserting fully expanded paths.

More annoying, there is some bug that is causing the box menu to be
truncated at 7 items. I tracked it down to happening in the initialization
before main() is called. I think it may be a gcc bug due to the huge
size of the Widget_Type.cxx source file. coding_panel.fl was not tried
due to this bug.


git-svn-id: file:///fltk/svn/trunk@5603 ea41ed52-d2ee-0310-a9c1-e6b18d33e121