More doxygen updates for Fl_Text_Buffer. Also changed some argument names from "char *s" to "char *text" to be in sync with the docs.
Caution: If there will be name conflicts with Fl_Text_buffer::text(),
then this can at least be fixed easier, because searching for "text"
is better than for "s". I remember that some argument names x,y,w,h
have been changed to uppercase, because of conflicts with x(), etc.
with ... valgrind (?) or maybe some compilers ?

I did this deliberately, because changing the headers and doxygen docs
to "s" instead of "text" appeared to be not feasible.


git-svn-id: file:///fltk/svn/branches/branch-1.3@6326 ea41ed52-d2ee-0310-a9c1-e6b18d33e121