Attempt to fix STR #2550 to make fl_text_extents work in Xlib-only (i.e. non-XFT) X11 builds.
This appears to be working now, though I can not test it fully, so we
need more feedback.
In particular the setting of the dx param is certainly wrong, but 
works fine in general cases. I think.
Also, I suspect it may behave badly in the face of RtoL text rendering
but I think we have other problems in that area anyway...



git-svn-id: file:///fltk/svn/branches/branch-1.3@8399 ea41ed52-d2ee-0310-a9c1-e6b18d33e121