fltk-1 compat layer broken after r9301.
OK, this gets the test1 folder building again on Win32, following on from the
fullscreen changes in r9301.

I think this fix is OK. Maybe. But I do not have a big enough range of
systems here to check, so *please* check this and correct any mistakes
I have made. Ta!




git-svn-id: file:///fltk/svn/branches/branch-3.0@9302 ea41ed52-d2ee-0310-a9c1-e6b18d33e121