Fix window background and label drawing ("plastic" scheme only).
There was a regression since svn r10598 because the background image width
and height were set to 0. The label drawing code erroneously used this
for positioning the label text inside the window which could be seen for
instance in test/valuators. Now the label (text) is not drawn at all.

PS: this should be fixed with a better scheme setup w/o manipulating
the window's label, labeltype, and image in a later version.


git-svn-id: http://seriss.com/public/fltk/fltk/branches/branch-1.3@10607 ea41ed52-d2ee-0310-a9c1-e6b18d33e121