Final commit to branch-1.3 *after* the release of FLTK 1.3.4.
See important notes below.

------------------------------------------------------------------------

Note to svn (subversion) users: This branch (branch-1.3) is finished
since the release of FLTK 1.3.4.

Only very serious bug fixes will be available in FLTK 1.3 if FLTK 1.4
is not yet released or not (yet) stable.

The new development branch is now branch-1.4 -- please checkout a new,
clean svn working copy if you want to get branch-1.4.

------------------------------------------------------------------------

Note to Git users: FLTK's svn branch-1.3 has been the 'master' branch
in our Git repository, but this branch is finished since the release
of FLTK 1.3.4.

Only very serious bug fixes will be available in FLTK 1.3 if FLTK 1.4
is not yet released or not (yet) stable.

The new development branch is now branch-1.4 -- please checkout
branch-1.4 if you want to get the current FLTK 1.4 development branch.

The Git 'master' branch will stay as-is for some time, but in the
near future branch-1.4 will likely become the new 'master' branch.

------------------------------------------------------------------------


git-svn-id: http://seriss.com/public/fltk/fltk/branches/branch-1.3@12100 ea41ed52-d2ee-0310-a9c1-e6b18d33e121