Windows: Disable dynamic linking/loading of winsock dll.
This *preliminary* commit disables dynamic linking/loading of winsock
(ws2_32.dll) and links it statically to fluid and all FLTK test
programs. This is done by conditional code (#if 0) that disables the
old code and #defines some macros. This *must* be removed and replaced
with the original socket interface once the code has proved to work as
expected.

Note: example programs in the examples/ folder are not yet modified.
Todo: check and fix examples/Makefile if necessary.


git-svn-id: http://seriss.com/public/fltk/fltk/branches/branch-1.4@12947 ea41ed52-d2ee-0310-a9c1-e6b18d33e121